refactor + responsive drag fixes
