+++ app/react/V2/Components/Layouts/PaneLayout/Pane.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { PaneProps } from './types';
+
+const Pane = ({ children, className, background = 'white' }: PaneProps) => (
+  <div style={{ background }} className={className}>
+    {children}
+  </div>
+);
+
+export { Pane };
