+++ app/react/V2/Components/Layouts/PaneLayout/types.ts
@@ -1,0 +1,11 @@
+type PaneProps = React.PropsWithChildren & {
+  background?: string;
+  className?: string;
+};
+
+type PaneLayoutProps = {
+  children: React.ReactElement<PaneProps>[];
+  className?: string;
+};
+
+export type { PaneProps, PaneLayoutProps };
